fix pretty json and workflow summary types error
