+++ src/components/pretty-json/pretty-json.types.tsx
@@ -1,3 +1,14 @@
 export type Props = {
-  json: object;
+  json: JsonValue;
 };
+
+export type JsonValue =
+  | string
+  | number
+  | boolean
+  | null
+  | JsonObject
+  | JsonArray;
+
+type JsonObject = { [key: string]: JsonValue };
+type JsonArray = JsonValue[];
