+++ src/views/workflow-summary-tab/workflow-summary-tab-json-view/workflow-summary-tab-json-view.types.ts
@@ -1,15 +1,6 @@
+import { JsonValue } from '@/components/pretty-json/pretty-json.types';
+
 export type Props = {
   inputJson: JsonValue;
   resultJson: JsonValue;
 };
-
-export type JsonValue =
-  | string
-  | number
-  | boolean
-  | null
-  | JsonObject
-  | JsonArray;
-
-type JsonObject = { [key: string]: JsonValue };
-type JsonArray = JsonValue[];
